requirement.txt
newspaper3k
news-fetch
flask
